Move player and lobby to own file
