+++ backend/models.py
@@ -1,14 +1,9 @@
 from __future__ import annotations
 
-import dataclasses
-from dataclasses import dataclass
 from enum import Enum, StrEnum, auto
 from typing import Annotated, Literal
-from uuid import uuid4
 
 from pydantic import BaseModel, Field
-
-from .channel import Channel
 
 
 class Role(Enum):
@@ -18,49 +13,6 @@ class Role(Enum):
     burger = auto()
     fry = auto()
     drink = auto()
-
-
-@dataclass
-class Lobby:
-    """
-    A class representing a game state.
-
-    Attributes:
-        id (str): The lobby's internal ID
-        players (dict[str, Player]): Map of player id to Players.
-        code (str): The code used to join the lobby
-        started (bool): Whether the game has started
-    """
-
-    players: dict[str, Player]
-    id: str = dataclasses.field(init=False, default_factory=lambda: uuid4().hex)
-    code: str = "ABC"
-    started: bool = False
-
-    def broadcast(self, msg: Message) -> None:
-        """Send a message to all players."""
-        for player in self.players.values():
-            player.send(msg)
-
-
-@dataclass
-class Player:
-    """
-    A class representing a player's ingame data.
-
-    Attributes:
-        id (str): The player's internal ID
-        role (Role): The player's role.
-        channel (Channel): Channel to send messages to frontend.
-    """
-
-    channel: Channel
-    role: Role
-    id: str = dataclasses.field(init=False, default_factory=lambda: uuid4().hex)
-
-    def send(self, msg: Message) -> None:
-        """Send a message to this player."""
-        self.channel.put(msg)
 
 
 class MessageKind(StrEnum):
